requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
celery>=5.3
redis>=5.0
django-redis>=5.4
Pillow>=10.0
python-magic>=0.4.27
